:truck: finder 페이지 api 파일 위치 변경 #69
