+++ src/pages/play/finder/Modal/SearchLobby/SearchModalContent.tsx
@@ -1,16 +1,14 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
+import { postEnterRoom } from '../../api';
 import * as styles from '../index.css';
+import MessageModalContent from '../MessageModalContent';
 
 import type { RequestEnterRoom } from '@/types/play';
 
 import ColoredButton from '@/components/button/ColoredButton';
 import Input from '@/components/input';
-
-import MessageModalContent from '@/pages/play/finder/Modal/MessageModalContent';
-
-import { postEnterRoom } from '@/services/finder';
 
 import { ROUTE } from '@/constants/routes';
