+++ src/pages/play/finder/index.ts
@@ -1,0 +1,2 @@
+export { default } from './page';
+export { getRoomList } from './api';
